+++ backend/main.py
@@ -1,12 +1,14 @@
 import os
-from datetime import datetime, timezone
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from models import MultProjetosInput, ProjetoDelete, ProjetoInput, ProjetoUpdate, ProjetosDelete
-from schema import Projeto, get_session, Session
-
+from routes.projects import router as projects_router
+from routes.users import router as users_router
 
 app = FastAPI(title="Gerenciamento de Projetos")
+
+app.include_router(projects_router, prefix="/projetos", tags=["Projetos"])
+app.include_router(users_router, prefix="/usuarios", tags=["Usuários"])
+
 allowed_origins = os.getenv("ALLOWED_ORIGINS", "").split(',')
 
 app.add_middleware(
@@ -18,179 +20,5 @@ app.add_middleware(
 )
 
 @app.get('/')
-async def root():
-    return {"CódigoCerto": "FullStack"}
-
-@app.get('/projetos')
-async def get_projetos(session: Session = Depends(get_session)):
-    projetos = session.query(Projeto).all()
-    
-    if projetos: 
-        return projetos
-    
-    return []
-
-@app.post('/projeto/criar')
-async def registrar_projeto(projeto_input: ProjetoInput, session: Session = Depends(get_session)):
-    if(not projeto_input.nome.strip()):
-       raise HTTPException(status_code=422, detail='O nome do projeto não pode ser vazio')
-    
-    if(projeto_input.prazo and projeto_input.prazo < datetime.now(timezone.utc)):
-        raise HTTPException(status_code=422, detail='O prazo já está vencido')
-    
-    projeto = Projeto(nome=projeto_input.nome.strip(), descricao=projeto_input.descricao, prazo=projeto_input.prazo)
-    session.add(projeto)
-    session.commit()
-    
-    # Retorna a lista de projetos atualizada
-    try:
-        projetos = session.query(Projeto).all()
-        return {
-            "projeto inserido": projeto,
-            "projetos": projetos
-            }
-    except:
-        
-        return {
-        "projeto inserido": projeto,
-        "projetos": []
-    }
-
-@app.post('/projetos/criar')
-async def registrar_multiplos_projetos(projetos_input: MultProjetosInput, session: Session = Depends(get_session)):
-    
-    projetos_inseridos = []
-    projetos_com_erro = []
-    
-    for projeto_input in projetos_input.projetos:
-        try:
-            if(not projeto_input.nome.strip()):
-                raise HTTPException(status_code=422, detail='Não pode haver nome de projeto vazio')
-            
-            projeto = Projeto(nome=projeto_input.nome.strip(), descricao=projeto_input.descricao, prazo=projeto_input.prazo)
-            session.add(projeto)
-            session.commit()
-            
-            projetos_inseridos.append({
-                "id": projeto.id,
-                "projeto": projeto.nome
-            })
-        except HTTPException as erro:
-            projetos_com_erro.append({
-                "projeto": projeto_input.nome,
-                "erro": str(erro.detail)
-            })
-
-    # Retorna a lista de projetos atualizada
-    try:
-        projetos = session.query(Projeto).all()
-        return {
-            "projetos inseridos": projetos_inseridos,
-            "projetos com erro": projetos_com_erro if projetos_com_erro else 0,
-            "projetos": projetos
-            }
-    except:
-        
-        return {
-        "projetos inseridos": projetos_inseridos,
-        "projetos com erro": projetos_com_erro if projetos_com_erro else 0,
-        "projetos": []
-    }
-
-@app.put('/projeto/editar')
-async def alterar_projeto(projeto_update: ProjetoUpdate, session: Session = Depends(get_session)):
-    projeto = session.query(Projeto).filter(projeto_update.id == Projeto.id).first()
-    
-    if not projeto:
-        raise HTTPException(status_code=404, detail='Projeto não encontrado')
-    
-    if projeto_update.nome.strip() != "":
-        projeto.nome = projeto_update.nome.strip()
-    if projeto_update.descricao != "":
-        projeto.descricao = projeto_update.descricao
-    if projeto_update.prazo != "":
-        projeto.prazo = projeto_update.prazo
-    
-    session.add(projeto)
-    session.commit()
-    
-    # Retorna a lista de projetos atualizada
-    try:
-        projetos = session.query(Projeto).all()
-
-        return {
-            "projeto atualizado":{
-                "id": projeto.id,
-                "nome": projeto.nome,
-                "descricao": projeto.descricao
-                },
-            "projetos": projetos
-            }
-    
-    except:
-        return {
-            "projeto atualizado":{
-                "id": projeto.id,
-                "nome": projeto.nome,
-                "descricao": projeto.descricao
-                },
-            "projetos": []
-            }
-
-    
-
-@app.delete('/projeto/deletar')
-async def remover_projeto(projeto_delete: ProjetoDelete, session: Session = Depends(get_session)):
-    projeto = session.query(Projeto).filter(projeto_delete.id == Projeto.id).first()
-    if not projeto:
-        raise HTTPException(status_code=404, detail='Projeto não encontrado')
-
-    session.delete(projeto)
-    session.commit()
-
-    # Retorna a lista de projetos atualizada
-    try:
-        projetos = session.query(Projeto).all()
-        return {"deletado": projeto.nome,
-                "projetos ": projetos}
-    except:
-        return {
-            "deletado": projeto.nome,
-            "projetos ": []
-            }
-
-@app.delete('/projetos/deletar')
-async def remover_multiplos_projetos(projeto_delete: ProjetosDelete, session: Session = Depends(get_session)):
-    projetos_deletados = []
-    projetos_nao_encontrados = []
-
-    for id in projeto_delete.ids:
-        try:
-            projeto = session.query(Projeto).filter(id == Projeto.id).first()
-            if not projeto:
-                raise HTTPException(status_code=404, detail='Projeto não encontrado')
-        
-            session.delete(projeto)
-            session.commit()
-            projetos_deletados.append(projeto.nome)
-
-        except HTTPException:
-            projetos_nao_encontrados.append(id)
-
-    # Retorna a lista de projetos atualizada
-    try:
-        projetos = session.query(Projeto).all()
-
-        return {
-            "projetos deletados": projetos_deletados,
-            "projetos não encontrados": projetos_nao_encontrados if projetos_nao_encontrados else [],
-            "projetos": projetos
-            }
-    
-    except:
-
-         return {
-             "projetos deletados": projetos_deletados,
-             "projetos não encontrados": projetos_nao_encontrados if projetos_nao_encontrados else [],
-             "projetos ": []
-             }
+async def index():
+    return {"Código Certo": "Full Stack"}
